Fix Array init, use single quotes, add tests
