+++ py-polars/polars/datatypes/_utils.py
@@ -33,14 +33,14 @@ def _dtype_to_init_repr_array(dtype: Array, prefix: str) -> str:
         inner_repr = dtype_to_init_repr(dtype.inner, prefix)
     else:
         inner_repr = ""
-    init_repr = f"{prefix}{class_name}({inner_repr}, width={dtype.width})"
+    init_repr = f"{prefix}{class_name}({inner_repr}, shape={dtype.shape})"
     return init_repr
 
 
 def _dtype_to_init_repr_struct(dtype: Struct, prefix: str) -> str:
     class_name = dtype.__class__.__name__
     inner_list = [
-        f'"{field_name}": {dtype_to_init_repr(inner_dtype, prefix)}'
+        f"{field_name!r}: {dtype_to_init_repr(inner_dtype, prefix)}"
         for field_name, inner_dtype in dict(dtype).items()
     ]
     inner_repr = "{" + ", ".join(inner_list) + "}"
